requirements.txt
Flask-Restless==0.16.0
# 0.17.0 is broken https://github.com/jfinkels/flask-restless/issues/409
Flask-JWT==0.3.2
Flask-Security==1.7.5
Flask-Admin==1.5.1
Flask-SQLAlchemy==2.3.2
sqlalchemy
